requirements.txt
Django==4.2.7
python-telegram-bot==13.7
djangorestframework==3.14.0
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
django-redis==6.0.0
